Reorganize and add test for get_themes
